charmander-api/charmander-api-tts/pyproject.toml
[project]
name = "charmander-api-tts"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
gtts = "^2.5.4"
boto3 = "^1.37.9"
python-dotenv = "^1.0.1"
nanoid = "^2.0.0"

[tool.poe.tasks.run]
help = "Run the application"
cmd = "uvicorn main:app --host 0.0.0.0 --port 8080"

[tool.poe.tasks.dev]
help = "Run the application for development"
cmd = "uvicorn main:app --reload"
